修改数据类型  questionNo short, sem boolean, year short

+++ src/com/company/dao/SubjectDao.java
@@ -5,6 +5,6 @@ import com.company.domain.Subject;
 public interface SubjectDao {
 
     public Subject findById(String id);
-    public Subject findByTeaClaId(int year,int sem,String teaId,String classId);
+    public Subject findByTeaClaId(short year,int sem,String teaId,String classId);
 
 }
