+++ src/com/company/dao/sqlplus/SetDaoImp.java
@@ -31,7 +31,7 @@ public class SetDaoImp implements SetDao {
                 result.setSem(rs.getBoolean("sem"));
                 result.setTea_ID(rs.getString("tea_id"));
                 result.setTestNo(rs.getString("test#"));
-                result.setYear(rs.getInt("year"));
+                result.setYear(rs.getShort("year"));
 
                 return result;
             }
@@ -66,14 +66,14 @@ public class SetDaoImp implements SetDao {
         return null;
     }
 
-    public void create(String testNo,int year,boolean sem,String teaId,String cId){
+    public void create(String testNo,short year,boolean sem,String teaId,String cId){
      try{
         Connection conn=TestApplication.conn;
         String sql="insert into set (test#,year,sem,tea_id,c_id) values (?,?,?,?,?)" ;
         PreparedStatement pstmt=conn.prepareStatement(sql);
 
         pstmt.setString(1,testNo);
-        pstmt.setInt(2,year);
+        pstmt.setShort(2,year);
         pstmt.setBoolean(3,sem);
         pstmt.setString(4,teaId);
         pstmt.setString(5,cId);
